Extract subcommand options from the driver file.
