+++ toolchain/driver/driver.h
@@ -8,7 +8,10 @@
 #include "common/command_line.h"
 #include "llvm/ADT/ArrayRef.h"
 #include "llvm/ADT/StringRef.h"
+#include "toolchain/driver/codegen_options.h"
+#include "toolchain/driver/compile_subcommand.h"
 #include "toolchain/driver/driver_env.h"
+#include "toolchain/driver/link_subcommand.h"
 
 namespace Carbon {
 
@@ -49,9 +52,6 @@ class Driver {
 
  private:
   struct Options;
-  struct CodegenOptions;
-  struct CompileOptions;
-  struct LinkOptions;
   class CompilationUnit;
 
   // Delegates to the command line library to parse the arguments and store the
